refactoring to add validation.py
