+++ tests/test_visualization.py
@@ -1,41 +1,12 @@
 import pytest
 import numpy as np
-from tinnsleep.visualization import (plotTimeSeries, assert_ax_equals_data,
+from tinnsleep.visualization import (plotTimeSeries,
                                      assert_x_labels_correct, zoom_effect,
                                      plotAnnotations)
 import matplotlib.pyplot as plt
 import numpy.testing as npt
 import mne
-
-
-def test_asserts_homemade():
-    """Check if homemade asserts catch the error"""
-    sfreq = 1
-    np.random.seed(42)
-    data = np.random.randn(400, 2)
-    fig, ax = plotTimeSeries(data)
-
-    # check if x-labels are incorrect values
-    with pytest.raises(AssertionError, match=r"Arrays are not equal\n\nMismatched elements: 399 / 400 (99.8%)*"):
-        # check if correct values
-        assert_ax_equals_data(data, ax, sfreq=2)
-
-    # check if y-label values are incorrect
-    with pytest.raises(AssertionError, match="labels are not the same"):
-        assert_x_labels_correct(data, ["0" for k in range(data.shape[1])])
-
-    np.random.seed(10)
-    data = np.random.randn(400, 2)
-
-    # check if y-label position are incorrect
-    with pytest.raises(AssertionError, match="Arrays are not equal"):
-        assert_x_labels_correct(data, ["0" for k in range(data.shape[1])])
-
-    # check if data values are incorrect values
-    with pytest.raises(AssertionError, match=r"Items are not equal to 7 significant digits:"):
-        # check if correct values
-        assert_ax_equals_data(data, ax, sfreq=sfreq)
-
+from tinnsleep.validation import assert_ax_equals_data
 
 def test_plotTimeSeries_noparams():
     """Complete test suite for plotTimeSeries
@@ -123,10 +94,10 @@ def test_plotTimeSeries_incorrect_parameters():
     np.random.seed(42)
     data = np.random.randn(400, 4)
 
-    with pytest.raises(ValueError, match="\`ch_names\` must be a list or an iterable of shape \(n_dimension,\) or None"):
+    with pytest.raises(ValueError, match="\`ch_names\` must be a list or an iterable of shape \(n_channels,\) or None"):
         plotTimeSeries(data, ch_names=True)
 
-    with pytest.raises(ValueError, match='ch_names should be same length as the number of channels of data'):
+    with pytest.raises(ValueError, match='`ch_names` should be same length as the number of channels of data'):
         plotTimeSeries(data, ch_names=[1, 2])
 
     with pytest.raises(ValueError, match="\`ax\` must be a matplotlib Axes instance or None"):
